chore: organize tests and tweak stream delays
