+++ tests/e2e/auth.setup.ts
@@ -1,9 +1,9 @@
-import path from 'path';
+import path from 'node:path';
 import { generateId } from 'ai';
 import { getUnixTime } from 'date-fns';
 import { expect, test as setup } from '@playwright/test';
 
-const authFile = path.join(__dirname, '../playwright/.auth/session.json');
+const authFile = path.join(__dirname, '../../playwright/.auth/session.json');
 
 setup('authenticate', async ({ page }) => {
   const testEmail = `test-${getUnixTime(new Date())}@playwright.com`;
