+++ tests/e2e/reasoning.setup.ts
@@ -1,10 +1,10 @@
-import path from 'path';
+import path from 'node:path';
 import { expect, test as setup } from '@playwright/test';
-import { ChatPage } from './pages/chat';
+import { ChatPage } from '../pages/chat';
 
 const reasoningFile = path.join(
   __dirname,
-  '../playwright/.reasoning/session.json',
+  '../../playwright/.reasoning/session.json',
 );
 
 setup('switch to reasoning model', async ({ page }) => {
